[IMP] purchase_sale_inter_company: add method _check_intercompany_product to purchase order line for better extensibility
