+++ purchase_sale_inter_company/README.rst
@@ -7,7 +7,7 @@ Inter Company Module for Purchase to Sale Order
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:e5062be0f49761f4653cc901fa95692b4cbd0f397e292bba54572234150b222f
+   !! source digest: sha256:8afec057f60529d15735ab62ab11ca70230885d40c35f8279df36b888066b8a0
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
