+++ purchase_sale_inter_company/__manifest__.py
@@ -6,7 +6,7 @@
 {
     "name": "Inter Company Module for Purchase to Sale Order",
     "summary": "Intercompany PO/SO rules",
-    "version": "16.0.1.1.0",
+    "version": "16.0.1.1.1",
     "category": "Purchase Management",
     "website": "https://github.com/OCA/multi-company",
     "author": "Odoo SA, Akretion, Tecnativa, Odoo Community Association (OCA)",
