+++ purchase_sale_inter_company/i18n/fr_FR.po
@@ -115,7 +115,7 @@ msgstr "Ligne de commande d'achat source"
 
 #. module: purchase_sale_inter_company
 #. odoo-python
-#: code:addons/purchase_sale_inter_company/models/purchase_order.py:0
+#: code:addons/purchase_sale_inter_company/models/purchase_order_line.py:0
 #, python-format
 msgid ""
 "The generated sale orders with reference %(orders)s can't be modified. "
@@ -155,7 +155,7 @@ msgstr ""
 
 #. module: purchase_sale_inter_company
 #. odoo-python
-#: code:addons/purchase_sale_inter_company/models/purchase_order.py:0
+#: code:addons/purchase_sale_inter_company/models/purchase_order_line.py:0
 #, python-format
 msgid ""
 "You can't change this purchase order as the corresponding sale is %(state)s"
@@ -163,7 +163,7 @@ msgstr ""
 
 #. module: purchase_sale_inter_company
 #. odoo-python
-#: code:addons/purchase_sale_inter_company/models/purchase_order.py:0
+#: code:addons/purchase_sale_inter_company/models/purchase_order_line.py:0
 #, python-format
 msgid "You cannot create SO from PO because product '%s' is not intercompany"
 msgstr ""
