+++ purchase_sale_inter_company/models/__init__.py
@@ -1,5 +1,7 @@
 from . import account_move
 from . import purchase_order
+from . import purchase_order_line
 from . import res_company
 from . import res_config
 from . import sale_order
+from . import sale_order_line
