+++ purchase_sale_inter_company/models/sale_order_line.py
@@ -1,0 +1,17 @@
+# Copyright 2013-Today Odoo SA
+# Copyright 2016-2019 Chafique DELLI @ Akretion
+# Copyright 2018-2019 Tecnativa - Carlos Dauden
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from odoo import fields, models
+
+
+class SaleOrderLine(models.Model):
+    _inherit = "sale.order.line"
+
+    auto_purchase_line_id = fields.Many2one(
+        comodel_name="purchase.order.line",
+        string="Source Purchase Order Line",
+        readonly=True,
+        copy=False,
+    )
